Introduces Driver\Pdo\AbstractPdo Introduces Laminas\Db\Adapter\Platform\PlatformInterface::getSqlPlatformDecorator() Improves typing around surface area methods Adds strict typing to all touched files.
