+++ src/Adapter/AdapterServiceFactory.php
