+++ src/Adapter/Driver/Pdo/_Feature/SqliteRowCounter.php
