+++ src/Sql/Platform/_Mysql/Ddl/AlterTableDecorator.php
